Refactor attendance history page
